requirements.txt
agno>=2.0.0
streamlit>=1.35.0
python-dotenv>=1.0.0
google-generativeai>=0.8.0
exa-py>=1.0.0
pydantic>=2.0.0
